Faults :Add support for double faults
Includes a test fixture to test a kernel stack overflow double
faults correctly. Previous to this change ,the kernel, upong a
kernel stack overflow, would throw a fault, attempt to push
exception parameters to the kernel stack, which is overflowing,
and would then triple fault and reset by default. We implement
stack switching, so that the expection stack frame is pushed onto a
seperate, hopefully not also overflowing, stack frame, and all is
right with the world.